refactor(input): extract useRecipeHub hook from InputEnvironmentPanel (#157)
Move recipe hub state and handlers (applyCuratedRecipe, handleFetchRemote,
handleImport, tab management) into a dedicated useRecipeHub hook. Reduces
the main component from 2,022 to 1,885 lines and lowers cyclomatic
complexity by extracting branching logic (try/catch, validation gates).
