+++ src/components/features/input/InputEnvironmentPanel.tsx
@@ -40,6 +40,7 @@ import { useHardwareProbe } from "@/lib/hooks/useHardwareProbe";
 import { navigatePipeline } from "@/lib/pipelineNavigation";
 import { estimateVramForCatalogPreset } from "@/lib/presetVramEstimate";
 import { presetDisplayName, useRecipeCatalog, type RecipeSortMode } from "@/components/features/input/useRecipeCatalog";
+import { useRecipeHub } from "@/components/features/input/useRecipeHub";
 import {
   getBaseName,
   formatFileSize,
@@ -126,166 +127,28 @@ export function InputEnvironmentPanel({
     handleClearToken,
   } = useHfToken();
 
-  // States for the Olive Recipe Hub
-  const [recipeSearch, setRecipeSearch] = useState("");
-  const [selectedArchitecture, setSelectedArchitecture] = useState<string>("All");
-  const [selectedDevice, setSelectedDevice] = useState<string>("All");
-  const [syncStatus, setSyncStatus] = useState<"idle" | "loading" | "success" | "error">("idle");
-  const [syncError, setSyncError] = useState("");
-  const [repoUrl, setRepoUrl] = useState(`https://github.com/${OLIVE_RECIPES_REPO}`);
-  const [repoBranch, setRepoBranch] = useState(OLIVE_RECIPES_BRANCH);
-  const [repoPath, setRepoPath] = useState(
-    "Qwen-Qwen2.5-1.5B-Instruct/NvTensorRtRtx/Qwen2.5-1.5B-Instruct_model_builder_fp16.json",
-  );
-  const [importJson, setImportJson] = useState("");
-  const [importError, setImportError] = useState<string | null>(null);
-  const [activeRecipeTab, setActiveRecipeTabRaw] = useState<"starter" | "github" | "editor">("starter");
-  const [visitedRecipeTabs, setVisitedRecipeTabs] = useState<Set<string>>(new Set(["starter"]));
-  const [, startRecipeTabTransition] = useTransition();
-  const setActiveRecipeTab = (tab: "starter" | "github" | "editor") => {
-    startRecipeTabTransition(() => {
-      setActiveRecipeTabRaw(tab);
-      setVisitedRecipeTabs((prev) => {
-        if (prev.has(tab)) return prev;
-        return new Set(prev).add(tab);
-      });
-    });
-  };
-  const [recipeSuccessMsg, setRecipeSuccessMsg] = useState<string | null>(null);
-  const [applyingRecipePath, setApplyingRecipePath] = useState<string | null>(null);
-  const [appliedRecipeLabel, setAppliedRecipeLabel] = useState<string | null>(null);
-  const [recipeRailExpanded, setRecipeRailExpanded] = useState(true);
-  const [sourceConfigExpanded, setSourceConfigExpanded] = useState(false);
+  // Recipe Hub (extracted to useRecipeHub)
+  const { data: hardwareProbe = null, isLoading: hardwareProbeLoading } = useHardwareProbe();
+  const {
+    recipeSearch, setRecipeSearch, selectedArchitecture, setSelectedArchitecture,
+    selectedDevice, setSelectedDevice, syncStatus, setSyncStatus, syncError, setSyncError,
+    repoUrl, setRepoUrl, repoBranch, setRepoBranch, repoPath, setRepoPath,
+    importJson, setImportJson, importError, setImportError,
+    activeRecipeTab, setActiveRecipeTab, visitedRecipeTabs,
+    recipeSuccessMsg, applyingRecipePath, appliedRecipeLabel,
+    recipeRailExpanded, setRecipeRailExpanded,
+    sourceConfigExpanded, setSourceConfigExpanded,
+    recipeRailCollapsed,
+    handleApplyCuratedRecipe, handleApplyCuratedRecipeAnyway,
+    handleFetchRemote, handleImport,
+  } = useRecipeHub({ setState, hardwareProbe });
+
   const [localModelHints, setLocalModelHints] = useState<LocalModelHints | null>(null);
   const [localHintsLoading, setLocalHintsLoading] = useState(false);
   const [showLocalRecipeMatchesOnly, setShowLocalRecipeMatchesOnly] = useState(false);
   const [hideIncompatibleRecipes, setHideIncompatibleRecipes] = useState(true);
-  const { data: hardwareProbe = null, isLoading: hardwareProbeLoading } = useHardwareProbe();
   const [recipeSort, setRecipeSort] = useState<RecipeSortMode>("recommended");
   const [expandedRecipeGroups, setExpandedRecipeGroups] = useState<Set<string>>(new Set());
-
-  const recipeRailCollapsed = Boolean(appliedRecipeLabel) && !recipeRailExpanded;
-
-  const applyCuratedRecipe = async (item: RecipeCatalogItem, options?: { allowIncompatible?: boolean }) => {
-    setApplyingRecipePath(item.repoPath);
-    setSyncStatus("idle");
-    setSyncError("");
-
-    try {
-      const json = await fetchOliveRecipesCatalogItem(item);
-      const metadata = compareCatalogMetadataToRecipe(item, json);
-      const hw = assessCatalogItemHardwareCompatibility(item, hardwareProbe, json);
-
-      if (hw.tier === "unavailable" && !options?.allowIncompatible) {
-        setSyncStatus("error");
-        setSyncError(
-          `Recipe targets ${hw.targetDevice} but this machine cannot run it. ${hw.reason} Use "Apply anyway" only for remote or cross-compile workflows.`,
-        );
-        return;
-      }
-
-      setState(deriveUiStateFromOliveRecipe(json, { replacePasses: true }));
-      setAppliedRecipeLabel(item.name);
-      setRecipeRailExpanded(false);
-      setSourceConfigExpanded(true);
-      setImportJson(JSON.stringify(json, null, 2));
-      setImportError(null);
-      const mismatchNote =
-        !metadata.matches && metadata.recipeDevice
-          ? ` Catalog device (${metadata.catalogDevice}) differs from recipe EP (${metadata.recipeDevice}).`
-          : "";
-      const approximateNote =
-        item.metadataSource !== "recipe" ? " Tags are folder-inferred (approximate)." : "";
-      const hwNote =
-        hw.tier === "unavailable"
-          ? " Applied despite missing local hardware (cross-compile / remote target)."
-          : hw.tier === "compatible"
-            ? ` Verified for ${hw.targetDevice} on this machine.`
-            : "";
-      setRecipeSuccessMsg(`Applied preset recipe: "${item.name}"!${approximateNote}${mismatchNote}${hwNote}`);
-      setTimeout(() => {
-        setRecipeSuccessMsg(null);
-      }, 5000);
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    } catch (err: any) {
-      setSyncStatus("error");
-      setSyncError(err.message || "Failed to load recipe from GitHub.");
-    } finally {
-      setApplyingRecipePath(null);
-    }
-  };
-
-  const handleApplyCuratedRecipe = (item: RecipeCatalogItem) => applyCuratedRecipe(item);
-  const handleApplyCuratedRecipeAnyway = (item: RecipeCatalogItem) =>
-    applyCuratedRecipe(item, { allowIncompatible: true });
-
-  const handleFetchRemote = async (overrides?: { url?: string; branch?: string; path?: string }) => {
-    const url = (overrides?.url ?? repoUrl).trim();
-    const branch = (overrides?.branch ?? repoBranch).trim() || "main";
-    const path = (overrides?.path ?? repoPath).trim();
-
-    if (overrides?.url !== undefined) setRepoUrl(overrides.url);
-    if (overrides?.branch !== undefined) setRepoBranch(overrides.branch);
-    if (overrides?.path !== undefined) setRepoPath(overrides.path);
-
-    if (!url) {
-      setSyncStatus("error");
-      setSyncError("GitHub repository URL is required.");
-      return;
-    }
-    if (!path) {
-      setSyncStatus("error");
-      setSyncError("Recipe path is required.");
-      return;
-    }
-
-    setSyncStatus("loading");
-    setSyncError("");
-
-    try {
-      const { json } = await fetchGitHubRecipeJson(url, branch, path);
-      setImportJson(JSON.stringify(json, null, 2));
-      setImportError(null);
-      setSyncStatus("success");
-      setRecipeSuccessMsg("Downloaded remote recipe payload! Inspect in Editor tab.");
-      setTimeout(() => setRecipeSuccessMsg(null), 4000);
-      setActiveRecipeTab("editor");
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    } catch (err: any) {
-      console.error(err);
-      setSyncStatus("error");
-      setSyncError(err.message || "Failed to download remote file. Check connection URL.");
-    }
-  };
-
-  const handleImport = (allowIncompatible = false) => {
-    const { recipe, schema } = parseRecipeJson(importJson);
-    if (!schema.valid) {
-      setImportError(`Recipe structure invalid:\n- ${schema.errors.join("\n- ")}`);
-      return;
-    }
-
-    const targetDevice = getCatalogDeviceFromRecipe(recipe) ?? "CPU";
-    const hw = assessRecipeHardwareCompatibility(targetDevice, hardwareProbe);
-    if (hw.tier === "unavailable" && !allowIncompatible) {
-      setImportError(
-        `Recipe targets ${hw.targetDevice} but this machine cannot run it.\n${hw.reason}\nUse "Apply anyway" for remote/cross-compile workflows.`,
-      );
-      return;
-    }
-
-    setState(deriveUiStateFromOliveRecipe(recipe, { replacePasses: true }));
-    setAppliedRecipeLabel("Custom JSON recipe");
-    setRecipeRailExpanded(false);
-    setSourceConfigExpanded(true);
-    setImportError(null);
-    setRecipeSuccessMsg(
-      hw.tier === "unavailable"
-        ? "Recipe applied (incompatible hardware — remote/cross-compile target)."
-        : "Recipe parsed and applied successfully!",
-    );
-    setTimeout(() => setRecipeSuccessMsg(null), 4000);
-  };
 
   const {
     filteredRecipes,
